+++ backend/bizyair/bizyair_extras/nodes_advanced_refluxcontrol.py
@@ -2,8 +2,8 @@
 Reference: https://github.com/kaibioinfo/ComfyUI_AdvancedRefluxControl
 """
 
-from bizyair.bizyair import BizyAirBaseNode, data_types
-from bizyair.bizyair.path_utils import path_manager as folder_paths
+from bizyair.core import BizyAirBaseNode, data_types
+from bizyair.core.path_utils import path_manager as folder_paths
 
 STRENGTHS = ["highest", "high", "medium", "low", "lowest"]
 STRENGTHS_VALUES = [1, 2, 3, 4, 5]
