+++ backend/bizyair/bizyair_extras/nodes_custom_sampler.py
@@ -1,4 +1,4 @@
-from bizyair.bizyair import BizyAirBaseNode, data_types
+from bizyair.core import BizyAirBaseNode, data_types
 
 
 class DisableNoise(BizyAirBaseNode):
