+++ backend/bizyair/bizyair_extras/nodes_dataset.py
@@ -1,4 +1,4 @@
-from bizyair.bizyair import BizyAirBaseNode
+from bizyair.core import BizyAirBaseNode
 
 
 class TrainDatasetAdd(BizyAirBaseNode):
