+++ backend/bizyair/bizyair_extras/nodes_differential_diffusion.py
@@ -1,4 +1,4 @@
-from bizyair.bizyair import BizyAirBaseNode, data_types
+from bizyair.core import BizyAirBaseNode, data_types
 
 
 class DifferentialDiffusion(BizyAirBaseNode):
