+++ backend/bizyair/bizyair_extras/nodes_janus_pro.py
@@ -1,4 +1,4 @@
-from bizyair.bizyair import BizyAirBaseNode
+from bizyair.core import BizyAirBaseNode
 
 
 class JanusModelLoader(BizyAirBaseNode):
