+++ backend/bizyair/bizyair_extras/nodes_sd3.py
@@ -1,6 +1,6 @@
 # sd3.5
-from bizyair.bizyair import BizyAirBaseNode, BizyAirNodeIO, data_types
-from bizyair.bizyair.path_utils import path_manager as folder_paths
+from bizyair.core import BizyAirBaseNode, BizyAirNodeIO, data_types
+from bizyair.core.path_utils import path_manager as folder_paths
 
 
 class TripleCLIPLoader(BizyAirBaseNode):
