+++ requirements.txt
@@ -1,1 +1,3 @@
+bizyair
+bizyui
 requests
